Adds create_project function to linter worker
Linter worker creates a new project on OpenShift when the lint proces
is completed successfully. Instead of having it in a separate file, it
helps to have the function in the same file.
